Added taskbar visibility
Refactored version of the original shapedwindow.rs example code, this version shows your application within the taskbar (as a workaround, not as an official method however it works fairly well).

- Sigma
